+++ apis/nucleus/src/sn/type.js
@@ -1,5 +1,6 @@
 import { generator as SNFactory } from '@nebula.js/supernova';
 import { satisfies } from 'semver';
+import extend from 'extend';
 import { load } from './load';
 
 /**
@@ -29,7 +30,7 @@ export default function create(info, halo, opts = {}) {
         stringified = JSON.stringify(sn.qae.properties.initial);
         return sn;
       }),
-    initialProperties(initial) {
+    initialProperties(initial, extendProperties = false) {
       return this.supernova().then(() => {
         const props = {
           qInfo: {
@@ -39,9 +40,15 @@ export default function create(info, halo, opts = {}) {
           version: type.version,
           showTitles: true,
           ...JSON.parse(stringified),
+        };
+        if (extendProperties) {
+          extend(true, props, initial);
+          return props;
+        }
+        return {
+          ...props,
           ...initial,
         };
-        return props;
       });
     },
   };