+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.2](https://github.com/qlik-oss/nebula.js/compare/v3.1.1...v3.1.2) (2023-01-30)
+
+### Bug Fixes
+
+- set container bg to white in serve ([#1061](https://github.com/qlik-oss/nebula.js/issues/1061)) ([1a62ad9](https://github.com/qlik-oss/nebula.js/commit/1a62ad9e24e983cfe30db90739be5d7517b43cfe))
+
+### Features
+
+- created object now can extend initial properties ([#1063](https://github.com/qlik-oss/nebula.js/issues/1063)) ([c14a1e6](https://github.com/qlik-oss/nebula.js/commit/c14a1e64a4a03d3aaef72949f01ec358d8ce2a22))
+- listbox popover api ([#1067](https://github.com/qlik-oss/nebula.js/issues/1067)) ([744fe21](https://github.com/qlik-oss/nebula.js/commit/744fe2189d3e0ade89b2855980c9e7d429765053))
+- **Listbox:** overflow disclaimer ([#1059](https://github.com/qlik-oss/nebula.js/issues/1059)) ([ed97b08](https://github.com/qlik-oss/nebula.js/commit/ed97b081729053262fa0b69c23509a5d9f3f63cd))
+
+### Reverts
+
+- Revert "fix!: created object now extend initial properties" ([6e07403](https://github.com/qlik-oss/nebula.js/commit/6e0740369ceaf9deefd15e5e560bbff499a1fcda))
+
 ## [3.1.1](https://github.com/qlik-oss/nebula.js/compare/v3.1.0...v3.1.1) (2023-01-22)
 
 ### Bug Fixes